Extract PublicationToolbar
